+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+* Send correct `repository` value to Bugsnag (mrnugget)
 * Fix timestamp in title attribute on deployment details page (mrnugget)
 * Fix linebreaks in Flowdock messages (mrnugget)
 * Fix deprecated import path of `uuid` package (Issue 1, mrnugget)